third_party/WebKit/LayoutTests/fast/css/variables/multiple-writes-to-inline-style.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id='testElem'></div>
<script>

test(function() {
  assert_equals(testElem.getAttribute('style'), null);
  testElem.style.setProperty('--foo', 'first');
  assert_equals(testElem.style.getPropertyValue('--foo'), 'first');
  assert_equals(getComputedStyle(testElem).getPropertyValue('--foo'), 'first');
  assert_equals(testElem.getAttribute('style'), '--foo:first;');
  testElem.style.setProperty('--foo', 'second');
  assert_equals(testElem.style.getPropertyValue('--foo'), 'second');
  assert_equals(testElem.getAttribute('style'), '--foo:second;');
}, "subsequent writes to inline style overwrite older values.")

</script>
